+++ packages/terra-navigation-prompt/src/terra-dev-site/doc/navigation-prompt/examples/NavigationPromptExample.jsx
@@ -10,6 +10,7 @@ const exampleTitle = 'Descriptive Notification Prompt Title';
 const exampleMessage = 'A Notification Prompt usually has an introductory warning instructing the user that there are unsaved changes or that there is a similar situation that warrants capturing the user\'s attention and requires that they take action before continuing.\n\nIt is good practice to include details about the originating source (page name, side-panel title, modal header title, etc.) and location description (form name, title from the section of the page, general position, etc.) to provide the user a contextual reference as to where they have an area needing attention: e.g. where data is about to be lost, the window about to be closed, the section about to be removed. Following the initial details, it is helpful to include a detailed description educating the user about any danger or caution as to what will happen to the current items about to be lost (form data, page content, etc.) if the user chooses each of the two actions provided in the notification prompt message.';
 const exampleRejectButtonText = 'Descriptive Reject Button Action';
 const exampleAcceptButtonText = 'Descriptive Accept Button Action';
+const exampleButtonOrder = 'acceptFirst';
 
 /**
  * The Input is a stateful component that renders a NavigationPrompt based upon its current input element value.
@@ -90,9 +91,10 @@ const Form = ({ title }) => {
             onClick={() => {
               inputCheckpointRef.current.resolvePrompts({
                 title: exampleTitle,
-                message: exampleMessage,
+                startMessage: exampleMessage,
                 rejectButtonText: exampleRejectButtonText,
                 acceptButtonText: exampleAcceptButtonText,
+                buttonOrder: exampleButtonOrder,
               }).then(() => {
                 setTimeStamp(Date.now());
               }).catch(() => {
@@ -135,9 +137,10 @@ const FormSwitcher = () => {
   function onSwitchForm(formKey) {
     formCheckpointRef.current.resolvePrompts({
       title: exampleTitle,
-      message: exampleMessage,
+      startMessage: exampleMessage,
       rejectButtonText: exampleRejectButtonText,
       acceptButtonText: exampleAcceptButtonText,
+      buttonOrder: exampleButtonOrder,
     }).then(() => {
       setActiveForm(formKey);
     }).catch(() => {