+++ packages/terra-notification-dialog/src/terra-dev-site/test/notification-dialog/CompleteNotificationDialog.test.jsx
@@ -1,52 +1,43 @@
-import React from 'react';
+import React, { useState } from 'react';
 import Button from 'terra-button';
 import NotificationDialog, { NotificationDialogVariants } from '../../../NotificationDialog';
 
-const clickOK = () => {
-  alert('You clicked OK'); // eslint-disable-line no-alert
+const clickConfirm = () => {
+  alert('You clicked confirm'); // eslint-disable-line no-alert
 };
 
-class CompleteNotificationDialog extends React.Component {
-  constructor() {
-    super();
+const CompleteNotificationDialog = () => {
+  const [isOpen, setIsOpen] = useState(false);
 
-    this.state = {
-      isOpen: false,
-    };
+  const handleOpenModal = () => {
+    setIsOpen(true);
+  };
 
-    this.handleOpenModal = this.handleOpenModal.bind(this);
-    this.handleCloseModal = this.handleCloseModal.bind(this);
-  }
+  const handleCloseModal = () => {
+    setIsOpen(false);
+  };
 
-  handleOpenModal() {
-    this.setState({ isOpen: true });
-  }
-
-  handleCloseModal() {
-    this.setState({ isOpen: false });
-  }
-
-  render() {
-    return (
-      <div>
-        <NotificationDialog
-          variant={NotificationDialogVariants.ALERT}
-          isOpen={this.state.isOpen}
-          title="Make sure that the title relates directly to the choices."
-          message="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
-          primaryAction={{
-            text: 'OK',
-            onClick: clickOK,
-          }}
-          secondaryAction={{
-            text: 'Close',
-            onClick: this.handleCloseModal,
-          }}
-        />
-        <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={this.handleOpenModal} />
-      </div>
-    );
-  }
-}
+  return (
+    <>
+      <NotificationDialog
+        variant={NotificationDialogVariants.ALERT}
+        isOpen={isOpen}
+        title="Make sure that the title relates directly to the choices."
+        startMessage="The Main Instruction is text used to provide more detail or define terminology. Don’t repeat the title verbatim."
+        acceptAction={{
+          text: 'Confirm',
+          onClick: clickConfirm,
+        }}
+        rejectAction={{
+          text: 'Close',
+          onClick: handleCloseModal,
+        }}
+        buttonOrder="acceptFirst"
+        emphasizedAction="accept"
+      />
+      <Button id="trigger-notification-dialog" text="Trigger NotificationDialog" onClick={handleOpenModal} />
+    </>
+  );
+};
 
 export default CompleteNotificationDialog;