+++ packages/terra-application-navigation/CHANGELOG.md
@@ -4,6 +4,16 @@ Changelog
 Unreleased
 ----------
 
+1.13.0 - (November 20, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+1.12.0 - (November 18, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 1.11.0 - (November 7, 2019)
 ------------------
 ### Added