+++ packages/terra-date-picker/CHANGELOG.md
@@ -9,6 +9,16 @@ Unreleased
 ### Changed
 * Updated color for selected dates and removed ignore axe rules.
 
+4.21.0 - (November 20, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
+4.20.0 - (November 18, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 4.19.0 - (November 7, 2019)
 ------------------
 ### Fixed