+++ packages/terra-infinite-list/CHANGELOG.md
@@ -4,6 +4,11 @@ Changelog
 Unreleased
 ----------
 
+3.22.0 - (November 18, 2019)
+------------------
+### Added
+* Added documentation for `key` on `Progressive Loading`.
+
 3.21.0 - (November 7, 2019)
 ------------------
 ### Changed