+++ packages/terra-tabs/CHANGELOG.md
@@ -4,6 +4,11 @@ ChangeLog
 Unreleased
 ----------
 
+6.19.0 - (November 20, 2019)
+------------------
+### Changed
+* Minor dependency updates
+
 6.18.0 - (November 7, 2019)
 ------------------
 ### Changed